binaryToDecimal() - can now convert signed binary numbers and binary numbers with a fractional part to decimal numbers.
isValidBinaryString() - can now validate binary strings with a decimal point.

+++ src/prog2/samcis/NumberConverter.java
@@ -1,7 +1,7 @@
 package prog2.samcis;
 
 public interface NumberConverter {
-    int binaryToDecimal(String binary) throws Exception;
+    double binaryToDecimal(String binary) throws Exception;
     int octalToDecimal(String octal);
     int hexadecimalToDecimal(String hexadecimal);
